--- a/ci/publish-crate.sh
+++ b/ci/publish-crate.sh
@@ -17,6 +17,7 @@ CRATES=(
   drone
   programs/native/{budget,bpf_loader,lua_loader,native_loader,noop,system,vote}
   .
+  fullnode
 )