--- a/fullnode/Cargo.toml
+++ b/fullnode/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+edition = "2018"
+name = "solana-fullnode"
+version = "0.1.0"
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "http://solana.com/"
+
+[dependencies]
+clap = "2.32.0"
+log = "0.4.2"
+rayon = "1.0.3"
+serde_json = "1.0.10"
+solana = { path = "..", version = "0.11.0" }
+solana-drone = { path = "../drone", version = "0.11.0" }
+solana-metrics = { path = "../metrics", version = "0.11.0" }
+solana-sdk = { path = "../sdk", version = "0.11.0" }
+
+[[bin]]
+name = "solana-fullnode"
+path = "src/fullnode.rs"