--- a/scripts/cargo-install-all.sh
+++ b/scripts/cargo-install-all.sh
@@ -9,3 +9,4 @@ set -x
 cargo install --path drone "$@"
 cargo install --path keygen "$@"
 cargo install --path . "$@"
+cargo install --path fullnode "$@"