### Old
```sh
mvines@sol:~/solana$ cargo build
    Finished dev [unoptimized + debuginfo] target(s) in 0.50s                                                                                                           
mvines@sol:~/solana$ touch src/bank.rs
mvines@sol:~/solana$ cargo build
   Compiling solana v0.11.0 (/Users/mvines/ws/s2)                                                                                                                       
    Finished dev [unoptimized + debuginfo] target(s) in 25.90s                                                                                                          
```
### New
```sh
mvines@sol:~/solana/fullnode$ cargo build
    Finished dev [unoptimized + debuginfo] target(s) in 0.35s                                                                                                 
mvines@sol:~/solana/fullnode$ touch ../src/bank.rs 
mvines@sol:~/solana/fullnode$ cargo build
   Compiling solana v0.11.0 (/Users/mvines/ws/solana)                                                                                                         
   Compiling solana-fullnode v0.1.0 (/Users/mvines/ws/solana/fullnode)                                                                                        
    Finished dev [unoptimized + debuginfo] target(s) in 18.03s 
```

Essentially when you only care about the fullnode program, touching files in `src/` will not trigger a rebuild of all other programs by default.  One could have still prevented the unnecessary rebuilding before with additional `cargo build` command-line args but I doubt most folks actually did this.  Now the default is probably want you want.

Just look at the last commit in this PR for now, as the PR contains a ton of other commits from other PRs that have yet to land.

If this looks good I'll do the same for all the other programs under `src/bin/...` (at which point `src/` ought to be turned into a crate called `libfullnode` or something, but that's for later later)