@mvines 2018-12-13
Rebased, CI is happy, and changes are minimal.    The rest of the programs in src/bin/ are next!